+++ train.py
@@ -8,7 +8,7 @@ from pathlib import Path
 import lightgbm as lgb
 from tensorboardX import SummaryWriter
 
-from callbacks import Statistics, TensorBoardCallback
+from callbacks import Statistics, TensorBoardCallback, ModelExtractionCallback
 from data_loader import train_data_loader
 from parameters import Parameters
 
@@ -36,59 +36,6 @@ writer = SummaryWriter(log_dir=log_path)
 # =============================================================================
 # 
 # =============================================================================
-class ModelExtractionCallback(object):
-    """
-    original author : momijiame
-    ref : https://blog.amedama.jp/entry/lightgbm-cv-model
-    description : Class for callback to extract trained models from lightgbm.cv(). 
-    note: This class depends on private class '_CVBooster', so there are some future risks. 
-    
-    usage:
-        extraction_cb = ModelExtractionCallback()
-        callbacks = [extraction_cb,]
-    
-        lgb.cv(params, dtrain, nfold=5, 
-               num_boost_round=9999,
-               early_stopping_rounds=EARLY_STOPPING_ROUNDS,
-               verbose_eval=verbose_eval,
-               callbacks=callbacks,
-               seed=0)
-        
-        models = extraction_cb.raw_boosters
-    
-    """
-
-    def __init__(self):
-        self._model = None
-
-    def __call__(self, env):
-        # _CVBooster の参照を保持する
-        self._model = env.model
-
-    def _assert_called_cb(self):
-        if self._model is None:
-            # コールバックが呼ばれていないときは例外にする
-            raise RuntimeError('callback has not called yet')
-
-    @property
-    def boosters_proxy(self):
-        self._assert_called_cb()
-        # Booster へのプロキシオブジェクトを返す
-        return self._model
-
-    @property
-    def raw_boosters(self):
-        self._assert_called_cb()
-        # Booster のリストを返す
-        return self._model.boosters
-
-    @property
-    def best_iteration(self):
-        self._assert_called_cb()
-        # Early stop したときの boosting round を返す
-        return self._model.best_iteration
-
-
 def handler(context):
     print(f'start training with parameters : {Parameters.as_dict()}, context : {context}')
